Set up payment submission in merlins connector
