+++ web/app/containers/checkout-payment/actions.js
@@ -1,15 +1,11 @@
 import {browserHistory} from 'progressive-web-sdk/dist/routing'
 import {createAction} from 'progressive-web-sdk/dist/utils/action-creation'
-import {makeRequest, makeJsonEncodedRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
-import {jqueryResponse} from 'progressive-web-sdk/dist/jquery-response'
-import checkoutPaymentParser from './checkout-payment-parser'
-
 import {getPaymentBillingFormValues} from '../../store/form/selectors'
-import {getEmailAddress, getCustomerEntityID} from '../../store/checkout/selectors'
+import {getEmailAddress} from '../../store/checkout/selectors'
 import {getShippingAddress} from '../../store/checkout/shipping/selectors'
-import {getIsLoggedIn} from '../app/selectors'
+import {splitFullName} from '../../utils/utils'
+import {submitPayment as submitPaymentCommand} from '../../integration-manager/checkout/commands'
 
-import {receiveCheckoutData} from '../../store/checkout/actions'
 
 export const receiveContents = createAction('Received CheckoutPayment Contents')
 export const toggleFixedPlaceOrder = createAction('Toggled the fixed "Place Order" container', ['isFixedPlaceOrderShown'])
@@ -18,86 +14,30 @@ export const toggleCompanyAptField = createAction('Showing the "Company" and "Ap
 export const toggleNewAddressFields = createAction('Toggled new address fields', ['newShippingAddressIsEnabled'])
 export const setCvvType = createAction('Setting CVV type', ['cvvType'])
 
-export const receiveResponse = (response) => {
-    return (dispatch) => {
-        return jqueryResponse(response)
-            .then(([$, $responseText]) => {
-                dispatch(receiveContents(checkoutPaymentParser($, $responseText)))
-            })
-    }
-}
+export const submitPayment = () => (dispatch, getState) => {
+    const currentState = getState()
+    const billingValues = getPaymentBillingFormValues(currentState)
+    const email = getEmailAddress(currentState)
+    const sameAddress = billingValues.billing_same_as_shipping
 
-export const fetchContents = () => {
-    return (dispatch) => {
-        return makeRequest(window.location.href)
-            .then((response) => dispatch(receiveResponse(response)))
-    }
-}
+    let address = {}
 
-export const submitPayment = () => {
-    return (dispatch, getState) => {
-        const currentState = getState()
-        const entityID = getCustomerEntityID(currentState)
-        const isLoggedIn = getIsLoggedIn(currentState)
-        const sameAddress = getPaymentBillingFormValues(currentState).billing_same_as_shipping
-
-        let address = {}
-        const email = getEmailAddress(currentState)
-
-        if (sameAddress) {
-            address = getShippingAddress(currentState).toJS()
-        } else {
-            const {
-                name,
-                company,
-                addressLine1,
-                addressLine2,
-                countryId,
-                city,
-                regionId,
-                postcode,
-            } = getPaymentBillingFormValues(currentState)
-            const names = name.split(' ')
-
-            address = {
-                firstname: names.slice(0, -1).join(' '),
-                lastname: names.slice(-1).join(' '),
-                company: company || '',
-                postcode,
-                city,
-                street: addressLine2 ? [addressLine1, addressLine2] : [addressLine1],
-                regionId,
-                countryId,
-                saveInAddressBook: false
-            }
+    if (sameAddress) {
+        address = getShippingAddress(currentState).toJS()
+    } else {
+        const {firstname, lastname} = splitFullName(billingValues.name)
+        address = {
+            ...billingValues,
+            firstname,
+            lastname,
+            email
         }
-
-        const paymentInformation = {
-            billingAddress: {
-                ...address
-            },
-            cartId: entityID,
-            email,
-            paymentMethod: {
-                additional_data: null,
-                method: 'checkmo',
-                po_number: null
-            }
-        }
-        const persistPaymentURL = `/rest/default/V1/${isLoggedIn ? 'carts/mine' : `guest-carts/${entityID}`}/payment-information`
-        // Save payment address for confirmation
-        dispatch(receiveCheckoutData({payment: {address}}))
-        makeJsonEncodedRequest(persistPaymentURL, paymentInformation, {method: 'POST'})
-            .then((response) => response.json())
-            .then((responseJSON) => {
-                // Looks like when it is successful, the responseJSON is a number
-                if (/^\d+$/.test(responseJSON)) {
-                    browserHistory.push({
-                        pathname: '/checkout/onepage/success/'
-                    })
-                } else {
-                    console.error(responseJSON.message)
-                }
-            })
     }
+
+    return dispatch(submitPaymentCommand(address))
+        .then((url) => {
+            browserHistory.push({
+                pathname: url
+            })
+        })
 }
